Edit gage chars and stats functional
Add functionality to buttons to add, edit and delete gage station characteristics and statistics. Currently there is a toaster.service error when new chars or stats are posted to NSS. To prevent these errors, the toaster.services are commented out in this commit. This bug will need to be addressed in the future.