Citation Modal edits, updated the single row editing function, added ID column
Last changes before pull request. The Citation Modal is used to select the citation for new chars/stats and editing chars/stats. There is a function in place to prevent multiple chars/stats from being edited at the same time. The Citation Modal had to be edited to prevent the last two columns from displaying when opened in the GagePage. Also, added an ID column to the Citation Modal when it is opened in the GagePage.