End of day progress 11/5
Station section is pretty near complete. Http post request function is in place, but the request it failing for some reason; probably to do with the isRegulated category.  Data in the table can be edited and the dropdown menus allow the user to select the data type for each of the data columns.