Reload citations after put/post/delete
Set the citations to reload after a http request. Also made the saveFilter() and requeryFilters() functions to only work when not opened in the gagepage. Commented out the setSelectedRegions() function in the Gagepage. Not sure what this was originally meant to accomplish, especially when the global selectedRegion is cleared when the Mainview page is opened.